updated main.js
I have added an if clause and then changed the old if to an else if and the old else if to an else. This is between lines 79 and 87 as before it was saying you were and then seemingly some random percentage away. See the image I have texted. This solves the problem.